show gaps in weather data as such
The default behaviour for Plotly line charts is to just connect
a data point with its neighbours. While this is absolutely fine
when having regular data, e.g. one data point for every hour over
the whole timespan covered by the plot, it is not so fine when
there are gaps of several hours or even days in the data. In that
case a straight line will connect a temperature value of e.g.
24 °C from noon two days ago with a value of 8 °C in the evening
of the current day. Those "jumps" are a rather odd sight for the
user, so it might be better to skip the connecion between those
two points and show a proper gap instead.