use mysql-connector-java from Maven repository and not from system
The Debian package libmysql-java was available in Ubuntu 18.04
and Debian 9, but it was removed in later versions of those
operating systems. Therefore, nobody with a somewhat recent
Debian-based will be able to use the system package out of the
box. Switching to the Maven repository as replacement is a good
solution which should work out of the box on most OSes nowadays.